// src/app/models/employee.model.ts
export interface Employee {
  employeeName: string;
  doneDreamsNumber: number;
  inProgressDreamsNumber: number;
  userName: string;
  email: string
  password: string;
  coniformPassword: string
  averageTime: number;
  pictureId?: string;
  availableBalance: number;
  transferredBalance: number;
}